Adding upstream version 0.99.12.

+++ src/scripts/22iso.sh
@@ -23,7 +23,9 @@ Iso ()
 		mkdir -p "${LIVE_ROOT}"/binary/casper
 		for MANIFEST in "${LIVE_ROOT}"/filesystem.manifest*
 		do
-			mv "${MANIFEST}" "${LIVE_ROOT}"/binary/casper/
+			if [ -e "${MANIFEST}" ]; then
+				mv "${MANIFEST}" "${LIVE_ROOT}"/binary/casper/
+			fi
 		done
 
 		# Remove indices
